RPC call to peer servers
The server needs some way of communicating to a peer group of servers.
In the current case, it will only talk to the backup server. But it is
likely that there will be more than 1 backup server in the future and it
needs to replicate the requests to more than one server.