Track request history instead of just ids
The requestId was tracked to inform the backup or primary server in case
it had already seen a request from the client or via replication from
the primary. However, the server responded with 'true' regardless of the
actual response to the request when it first handled the request.

This is now fixed. The server will track the response history based on
the request id. As a result, it will not misinform the client and will
return the same response it had when it first handled the request.