Server handler
Lay the foundation to create a server handler independent of the view
server. The goal is to make the view server the dispatch layer. The
Handler will expose all the methods that should be exposed via RPC.

[a] - New ViewServerHandler to handle Get and Ping requests for the
ViewServer
[a] - dispatch with test Ping and test Get

TODO - move handler logic from ViewServer to ViewServerHanlder
TODO - change the return values for test Ping and Get to ensure that an
empty response is not being received